API

API

System

CollectiveSpins.system.Spin — Type.

A class representing a single spin.

A spin is defined by its position and its detuning to a main frequency.

Arguments

position: A vector defining a point in R3.

delta: Detuning.

source

CollectiveSpins.system.SpinCollection — Type.

A class representing a system consisting of many spins.

Arguments

spins: Vector of spins.

polarization: Unit vector defining the polarization axis.

gamma: Decay rate. (Has to be the same for all spins.)

source

CollectiveSpins.system.CavityMode — Type.

A class representing a single mode in a cavity.

Arguments

cutoff: Number of included Fock states.

delta=0 Detuning.

eta=0: Pump strength.

kappa=0: Decay rate.

source

CollectiveSpins.system.CavitySpinCollection — Type.

A class representing a cavity coupled to many spins.

Arguments

cavity: A CavityMode.

spincollection: A SpinCollection.

g: A vector defing the coupling strengths between the i-th spin and the cavity mode. Alternatively a single number can be given for identical coupling for all spins.

source

Geometry

CollectiveSpins.geometry.chain — Function.

geometry.chain(a, N)

Positions of spins on a chain in x-direction.

The chain starts at the origin and continues into positive x-direction.

Arguments

a: Spin-spin distance.

N: Number of spins

source

CollectiveSpins.geometry.triangle — Function.

geometry.triangle(a)

Positions of spins on a equilateral triangle in the xy-plane with edge length a.

The positions are: (0,0,0), (a,0,0), (a/2, h, 0)

source

CollectiveSpins.geometry.rectangle — Function.

geometry.rectangle(a, b; Nx=2, Ny=2)

Positions of spins on a rectangular lattice in the xy-plane.

The lattice starts at the origin and continues into positive x and y direction.

Arguments

a: Spin-spin distance in x-direction.

b: Spin-spin distance in y-direction.

Nx=2: Number of spins into x direction.

Ny=2: Number of spins into y direction.

source

CollectiveSpins.geometry.square — Function.

geometry.square(a; Nx=2, Ny=2)

Positions of spins on a square lattice in the xy-plane with distance a.

The lattice starts at the origin and continues into positive x and y direction.

source

CollectiveSpins.geometry.hexagonal — Function.

geometry.hexagonal(a; Nr=1)

Positions of spins on a hexagonal lattice in the xy-plane.

The hexagonal lattice consists of Nr rings.

source

CollectiveSpins.geometry.box — Function.

geometry.box(a, b, c; Nx=2, Ny=2, Nz=2)

Positions of spins on a orthorhombic lattice.

The lattice starts at the origin and continues into positive x, y and z direction.

source

CollectiveSpins.geometry.cube — Function.

geometry.cube(a; Nx=2, Ny=2, Nz=2)

Positions of spins on a cubic lattice with edge length a.

The lattice starts at the origin and continues into positive x, y and z direction.

source

Dipole-Dipole Interaction

CollectiveSpins.interaction.Omega — Function.

interaction.Omega(a, Θ, γ)

Dipole-dipole interaction frequency.

Calculates

\[Ω(a, θ, γ) = \frac{3}{4}γ G_θ(2π a)\]

Arguments

a: Distance between dipoles normalized by transition wavelength.

θ: Angle between the line connecting the two dipoles and the polarization axis.

γ: Single spin decay rate.

source

interaction.Omega(xi, xj, e, γ)

Dipole-dipole interaction frequency.

Calculates

\[Ω(|x_j-x_i|, θ, γ) = \frac{3}{4}γ G_θ(2π |x_j - x_i|)\]

with $θ = ∠(x_j-x_i, e)$.

Arguments

xi: Position of first spin.

xj: Position of second spin.

e: Polarization axis.

γ: Single spin decay rate.

source

CollectiveSpins.interaction.Gamma — Function.

interaction.Gamma(a, θ, γ)

Collective decay rate.

Calculates

\[Γ(a, θ, γ) = \frac{3}{2}γ F_θ(2π a)\]

Arguments

a: Distance between dipoles normalized by transition wavelength.

θ: Angle between the line connecting the two dipoles and the polarization axis.

γ: Single spin decay rate.

source

interaction.Gamma(xi, xj, e, γ)

Collective decay rate.

Calculates

\[Γ(|x_j-x_i|, θ, γ) = \frac{3}{2}γ F_θ(2π |x_j - x_i|)\]

with $θ = ∠(x_j-x_i, e)$.

Arguments

xi: Position of first spin.

xj: Position of second spin.

e: Polarization axis.

γ: Single spin decay rate.

source

CollectiveSpins.interaction.GammaMatrix — Function.

interaction.GammaMatrix(S::SpinCollection)

Matrix of the collective decay rate for a given SpinCollection.

source

CollectiveSpins.interaction.OmegaMatrix — Function.

interaction.OmegaMatrix(S::SpinCollection)

Matrix of the dipole-dipole interaction for a given SpinCollection.

source

Effective Interactions

CollectiveSpins.effective_interaction.triangle_orthogonal — Function.

effective_interaction.triangle_orthogonal(a)

Effective Omega and Gamma for a equilateral triangle with edge length a.

The polarization axis is orthogonal to the triangle plane.

source

CollectiveSpins.effective_interaction.square_orthogonal — Function.

effective_interaction.square_orthogonal(a)

Effective Omega and Gamma for a square with edge length a.

The polarization axis is orthogonal to the square plane.

source

CollectiveSpins.effective_interaction.rectangle_orthogonal — Function.

effective_interaction.rectangle_orthogonal(a, b)

Effective Omega and Gamma for a rectangle with edge lengths a and b.

The polarization axis is orthogonal to the rectangle plane.

source

CollectiveSpins.effective_interaction.cube_orthogonal — Function.

effective_interaction.cube_orthogonal(a)

Effective Omega and Gamma for a cube with edge length a

The polarization axis is orthogonal to the xy faces.

source

CollectiveSpins.effective_interaction.box_orthogonal — Function.

effective_interaction.box_orthogonal(a, b, c)

Effective Omega and Gamma for a box with edge lengths a, b and c.

The polarization axis is orthogonal to the top face.

source

CollectiveSpins.effective_interaction.chain — Function.

effective_interaction.chain(a, Θ, N)

Effective Omega and Gamma for an infinite chain.

The calculation is done by adding N spins left and N spins right of a central spin.

Arguments

a: Spin-spin distance.

θ: Angle between polarization axis and spin chain.

N: Number of included spins.

source

CollectiveSpins.effective_interaction.chain_orthogonal — Function.

effective_interaction.chain_orthogonal(a, N)

Effective Omega and Gamma for an infinite chain with orthogonal polarization axis.

The calculation is done by adding N spins left and N spins right of a central spin.

Arguments

a: Spin-spin distance.

N: Number of included spins.

source

CollectiveSpins.effective_interaction.squarelattice_orthogonal — Function.

effective_interaction.squarelattice_orthogonal(a, N)

Effective Omega and Gamma for a infinite square lattice.

The polarization axis is orthogonal to the square lattice plane and the calculation is done by creating a (2N+1)*(2N+1) square lattice and calculate the combined interaction for the central spin.

Arguments

a: Spin-spin distance.

N: Number of included spins.

source

CollectiveSpins.effective_interaction.hexagonallattice_orthogonal — Function.

effective_interaction.hexagonallattice_orthogonal(a, N)

Effective Omega and Gamma for a infinite hexagonal lattice.

The polarization axis is orthogonal to the square lattice plane and the calculation is done by creating hexagonal lattice consisting of N rings and calculate the combined interaction for the central spin.

Arguments

a: Spin-spin distance.

N: Number of included spins.

source

CollectiveSpins.effective_interaction.cubiclattice_orthogonal — Function.

effective_interaction.cubiclattice_orthogonal(a, N)

Effective Omega and Gamma for a infinite cubic lattice.

The polarization axis is orthogonal to the top face of a unit cell and the calculation is done by creating a (2N+1)(2N+1)(2N+1) cubic lattice and calculate the combined interaction for the central spin.

Arguments

a: Spin-spin distance.

N: Number of included spins.

source

CollectiveSpins.effective_interaction.tetragonallattice_orthogonal — Function.

effective_interaction.tetragonallattice_orthogonal(a, b, N)

Effective Omega and Gamma for a infinite tetragonal lattice.

The polarization axis is orthogonal to the top face of a unit cell and the calculation is done by creating a (2N+1)(2N+1)(2N+1) tetragonal lattice and calculate the combined interaction for the central spin.

Arguments

a: Spin-spin distance for bottom side square.

b: Height of the unit cell.

N: Number of included spins.

source

CollectiveSpins.effective_interaction.hexagonallattice3d_orthogonal — Function.

effective_interaction.hexagonallattice3d_orthogonal(a, b, N)

Effective Omega and Gamma for a infinite 3D hexagonal lattice.

The lattice consists of stacked planes of hexagonal lattices where the the polarization axis is orthogonal to the planes. The calculation is done by creating hexagonal lattices with N rings, stacking 2N+1 lattices of this kind above each other and calculating the combined interaction for the central spin.

Arguments

a: Spin-spin distance for hexagons.

b: Distance between planes of hexagonal lattices

N: Number of included spins.

source

Rotated effective interactions

CollectiveSpins.effective_interaction_rotated.square_orthogonal — Function.

effective_interaction_rotated.square_orthogonal(a, Nδ)

Effective Omega and Gamma for a square.

The polarization axis is orthogonal to the square plane.

Arguments

a: Edge length.

Nδ: Phase shift (Number of atoms in 2π).

source

CollectiveSpins.effective_interaction_rotated.cube_orthogonal — Function.

effective_interaction_rotated.cube_orthogonal(a, dϕ)

Effective Omega and Gamma for a cube.

The polarization axis is orthogonal to the xy faces.

Arguments

a: edge length.

dϕ: Phase shift.

source

CollectiveSpins.effective_interaction_rotated.chain_orthogonal — Function.

effective_interaction_rotated.chain_orthogonal(a, N, dϕ)

Effective Omega and Gamma for an infinite chain.

The polarization axis is orthogonal to the chain and the calculation is done by adding N spins left and N spins right of a central spin.

Arguments

a: Spin-spin distance.

N: Number of included spins.

dϕ: Phase shift between neighboring spins.

source

Methods

Quantum

CollectiveSpins.quantum.basis — Function.

quantum.basis(x)

Get basis of the given System.

source

CollectiveSpins.quantum.blochstate — Function.

quantum.blochstate(phi, theta[, N=1])

Product state of N single spin Bloch states.

All spins have the same azimuthal angle phi and polar angle theta.

source

CollectiveSpins.quantum.dim — Function.

quantum.dim(state)

Number of spins described by this state.

source

CollectiveSpins.quantum.Hamiltonian — Function.

quantum.Hamiltonian(S)

Hamiltonian of the given System.

source

CollectiveSpins.quantum.JumpOperators — Function.

quantum.JumpOperators(S)

Jump operators of the given system.

source

CollectiveSpins.quantum.JumpOperators_diagonal — Function.

quantum.JumpOperators_diagonal(S)

Jump operators of the given system. (diagonalized)

Diagonalized means that the Gamma matrix is diagonalized and the jump operators are changed accordingly.

source

CollectiveSpins.quantum.timeevolution_diagonal — Function.

quantum.timeevolution_diagonal(T, S, state0[; fout])

Master equation time evolution. (diagonalized)

Diagonalized means that the Gamma matrix is diagonalized and the jump operators are changed accordingly.

Arguments

T: Points of time for which output will be generated.

S: System

ρ₀: Initial density operator.

fout (optional): Function with signature fout(t, state) that is called whenever output should be generated.

source

CollectiveSpins.quantum.timeevolution — Function.

quantum.timeevolution(T, S, state0[; fout])

Master equation time evolution.

Diagonalized means that the Gamma matrix is diagonalized and the jump operators are changed accordingly.

Arguments

T: Points of time for which output will be generated.

S: System

ρ₀: Initial density operator.

fout (optional): Function with signature fout(t, state) that is called whenever output should be generated.

source

CollectiveSpins.quantum.rotate — Function.

meanfield.rotate(axis, angles, state)

Rotations on the Bloch sphere for the given density operator.

Arguments

axis: Rotation axis.

angles: Rotation angle(s).

ρ: Density operator that should be rotated.

source

CollectiveSpins.quantum.squeeze — Function.

quantum.squeeze(axis, χT, ρ₀)

Spin squeezing along an arbitrary axis.

Arguments

axis: Squeezing axis.

χT: Squeezing strength.

ρ₀: Operator that should be squeezed.

source

CollectiveSpins.quantum.squeezingparameter — Function.

quantum.squeezingparameter(ρ)

Calculate squeezing parameter for the given state.

source

0th order: Independent spins

CollectiveSpins.independent.blochstate — Function.

independent.blochstate(phi, theta[, N=1])

Product state of N single spin Bloch states.

All spins have the same azimuthal angle phi and polar angle theta.

source

CollectiveSpins.independent.dim — Function.

independent.dim(state)

Number of spins described by this state.

source

CollectiveSpins.independent.splitstate — Function.

independent.splitstate(state)

Split state into sx, sy and sz parts.

source

CollectiveSpins.independent.densityoperator — Function.

independent.densityoperator(sx, sy, sz)
independent.densityoperator(state)

Create density operator from independent sigma expectation values.

source

CollectiveSpins.independent.sx — Function.

independent.sx(state)

Sigma x expectation values of state.

source

CollectiveSpins.independent.sy — Function.

independent.sy(state)

Sigma y expectation values of state.

source

CollectiveSpins.independent.sz — Function.

independent.sz(state)

Sigma z expectation values of state.

source

CollectiveSpins.independent.timeevolution — Function.

independent.timeevolution(T, gamma, state0)

Independent time evolution.

Arguments

T: Points of time for which output will be generated.

gamma: Single spin decay rate.

state0: Initial state.

source

independent.timeevolution(T, S::SpinCollection, state0)

Independent time evolution.

Arguments

T: Points of time for which output will be generated.

S: SpinCollection describing the system.

state0: Initial state.

source

1st order: Meanfield

CollectiveSpins.meanfield.ProductState — Type.

Class describing a Meanfield state (Product state).

The data layout is [sx1 sx2 ... sy1 sy2 ... sz1 sz2 ...]

Arguments

N: Number of spins.

data: Vector of length 3*N.

source

CollectiveSpins.meanfield.blochstate — Function.

meanfield.blochstate(phi, theta[, N=1])

Product state of N single spin Bloch states.

All spins have the same azimuthal angle phi and polar angle theta.

source

CollectiveSpins.meanfield.dim — Function.

meanfield.dim(state)

Number of spins described by this state.

source

CollectiveSpins.meanfield.splitstate — Function.

meanfield.splitstate(N, data)
meanfield.splitstate(state)

Split state into sx, sy and sz parts.

source

CollectiveSpins.meanfield.densityoperator — Function.

mpc.densityoperator(state)

Create density operator from MPCState.

source

meanfield.densityoperator(sx, sy, sz)
meanfield.densityoperator(state)

Create density operator from independent sigma expectation values.

source

CollectiveSpins.meanfield.sx — Function.

meanfield.sx(state)

Sigma x expectation values of state.

source

CollectiveSpins.meanfield.sy — Function.

meanfield.sy(state)

Sigma y expectation values of state.

source

CollectiveSpins.meanfield.sz — Function.

meanfield.sz(state)

Sigma z expectation values of state.

source

CollectiveSpins.meanfield.timeevolution — Function.

meanfield.timeevolution(T, S::SpinCollection, state0[; fout])

Meanfield time evolution.

Arguments

T: Points of time for which output will be generated.

S: SpinCollection describing the system.

state0: Initial ProductState.

fout (optional): Function with signature fout(t, state) that is called whenever output should be generated.

source

CollectiveSpins.meanfield.timeevolution_symmetric — Function.

meanfield.timeevolution_symmetric(T, state0, Ωeff, Γeff[; γ, δ0, fout])

Symmetric meanfield time evolution.

Arguments

T: Points of time for which output will be generated.

state0: Initial ProductState.

Ωeff: Effective dipole-dipole interaction.

Γeff: Effective collective decay rate.

γ=1: Single spin decay rate.

δ0=0: Phase shift for rotated symmetric meanfield time evolution.

fout (optional): Function with signature fout(t, state) that is called whenever output should be generated.

source

CollectiveSpins.meanfield.rotate — Function.

meanfield.rotate(axis, angles, state)

Rotations on the Bloch sphere for the given ProductState.

Arguments

axis: Rotation axis.

angles: Rotation angle(s).

state: ProductState that should be rotated.

source

2nd order: Meanfield plus Correlations (MPC)

CollectiveSpins.mpc.MPCState — Type.

Class describing a MPC state (Product state + Correlations).

The data layout is vector that in matrix form looks like

Cxx Cxy Cyy Cxz Czz Cyz

where the Cij are the appropriate correlation matrices. The expectation values sx, sy and sz are the diagonals of the matrices Cxx, Cyy and Czz, respectively.

Arguments

N: Number of spins.

data: Vector of length (3N)(2*N+1).

source

CollectiveSpins.mpc.blochstate — Function.

mpc.blochstate(phi, theta[, N=1])

Product state of N single spin Bloch states.

All spins have the same azimuthal angle phi and polar angle theta.

source

CollectiveSpins.mpc.dim — Function.

mpc.dim(state)

Number of spins described by this state.

source

CollectiveSpins.mpc.splitstate — Function.

mpc.splitstate(N, data)
mpc.splitstate(state)

Returns sx, sy, sz, Cxx, Cyy, Czz, Cxy, Cxz, Cyz.

source

CollectiveSpins.mpc.correlation2covariance — Function.

mpc.correlation2covariance(corstate)

Convert a MPCState from correlation form into covariance form.

Basically it just calculates Cov_ab = <s_a s_b> - <s_a> <s_b>.

source

CollectiveSpins.mpc.covariance2correlation — Function.

mpc.covariance2correlation(covstate)

Convert a MPCState from covariance form into correlation form.

Basically it just calculates <s_a s_b> = Cov_ab + <s_a> <s_b>.

source

CollectiveSpins.meanfield.densityoperator — Method.

mpc.densityoperator(state)

Create density operator from MPCState.

source

CollectiveSpins.mpc.sx — Function.

mpc.sx(state)

Sigma x expectation values of state.

source

CollectiveSpins.mpc.sy — Function.

mpc.sy(state)

Sigma y expectation values of state.

source

CollectiveSpins.mpc.sz — Function.

mpc.sz(state)

Sigma z expectation values of state.

source

CollectiveSpins.mpc.Cxx — Function.

mpc.Cxx(state)

Sigmax-Sigmax correlation values of MPCState.

source

CollectiveSpins.mpc.Cyy — Function.

mpc.Cyy(state)

Sigmay-Sigmay correlation values of MPCState.

source

CollectiveSpins.mpc.Czz — Function.

mpc.Czz(state)

Sigmaz-Sigmaz correlation values of MPCState.

source

CollectiveSpins.mpc.Cxy — Function.

mpc.Cxy(state)

Sigmax-Sigmay correlation values of MPCState.

source

CollectiveSpins.mpc.Cxz — Function.

mpc.Cxz(state)

Sigmax-Sigmaz correlation values of MPCState.

source

CollectiveSpins.mpc.Cyz — Function.

mpc.Cyz(state)

Sigmay-Sigmaz correlation values of MPCState.

source

CollectiveSpins.mpc.timeevolution — Function.

mpc.timeevolution(T, S::SpinCollection, state0[; fout])

MPC time evolution.

Arguments

T: Points of time for which output will be generated.

S: SpinCollection describing the system.

state0: Initial MPCState.

fout (optional): Function with signature fout(t, state) that is called whenever output should be generated.

source

CollectiveSpins.mpc.rotate — Function.

mpc.rotate(axis, angles, state)

Rotations on the Bloch sphere for the given MPCState.

Arguments

axis: Rotation axis.

angles: Rotation angle(s).

state: MPCState that should be rotated.

source

CollectiveSpins.mpc.var_Sx — Function.

mpc.var_Sx(state0)

Variance of the total Sx operator for the given MPCState.

source

CollectiveSpins.mpc.var_Sy — Function.

mpc.var_Sy(state)

Variance of the total Sy operator for the given MPCState.

source

CollectiveSpins.mpc.var_Sz — Function.

mpc.var_Sz(state)

Variance of the total Sz operator for the given MPCState.

source

CollectiveSpins.mpc.squeeze — Function.

mpc.squeeze(axis, χT, state0)

Spin squeezing along an arbitrary axis.

Arguments

axis: Squeezing axis.

χT: Squeezing strength.

state0: MPCState that should be squeezed.

source

CollectiveSpins.mpc.squeezingparameter — Function.

mpc.squeezing_parameter(state)

Calculate squeezing parameter for the given state.

source